+++ NEWS.md
@@ -5,6 +5,7 @@
 * 35 palettes have been added from MoMAColors package. (#142)
 * 32 palettes have been added from PrettyCols package. (#126)
 * 15 palettes have been added from waRhol package. (#131)
+* 18 palettes have been added from MexBrewer package. (#137)
 
 # paletteer 1.5.0